+++ src/main/java/my/project/services/SupplierService.java
@@ -1,11 +1,14 @@
 package my.project.services;
 
 import my.project.entities.abm.CategoryProduct;
+import my.project.entities.abm.Product;
 import my.project.entities.abm.Supplier;
+import my.project.repository.ProductRepository;
 import my.project.repository.SupplierRepository;
 import my.project.services.Interface.InAbmService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -16,18 +19,21 @@ public class SupplierService implements InAbmService<Supplier, Integer> {
 
     @Autowired
     private SupplierRepository supplierRepository;
+    @Autowired
+    private ProductRepository productRepository;
 
-    @Override
     public ResponseEntity<Supplier> create(Supplier entity) {
         return ResponseEntity.ok(supplierRepository.save(entity));
     }
 
-    @Override
     public ResponseEntity<List<Supplier>> findAll() {
         return ResponseEntity.ok(supplierRepository.findAll());
     }
 
-    @Override
+    public ResponseEntity<List<Supplier>> findByFilter(String filter) {
+        return ResponseEntity.ok(supplierRepository.findByFilter(filter));
+    }
+
     public ResponseEntity<Supplier> update(Integer id, Supplier entity) {
         Supplier supplierFound = supplierRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("category not found"));
@@ -36,8 +42,9 @@ public class SupplierService implements InAbmService<Supplier, Integer> {
             supplierFound.setName(entity.getName());
             supplierFound.setTelephone(entity.getTelephone());
             supplierFound.setRuc(entity.getRuc());
+            supplierFound.setAddress(entity.getAddress());
 
-            return ResponseEntity.ok(supplierRepository.save(entity));
+            return ResponseEntity.ok(supplierRepository.save(supplierFound));
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
@@ -46,6 +53,12 @@ public class SupplierService implements InAbmService<Supplier, Integer> {
     public ResponseEntity<String> delete(Integer id) {
         Supplier supplierFound = supplierRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("category not found"));
+
+        List<Product> productList = productRepository.findBySupplier(id);
+
+        if(productList.size() > 0){
+            return new ResponseEntity<>("not found", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
 
         supplierRepository.delete(supplierFound);
